Added support for GWT as an 'OUTPUT_LANGUAGE'. This is really a modified version of the Java output language but without dependencies on IO classes that are not emulated by the GWT JRE emulation (as the output will be converted in JavaScript).
As part of this commit, I have added catch-all statements to if statements that branch around output languages (added a final else to catch an unexpected language name). Previously, there appeared to be an assumption of either Java or C++, even though other language names could be supplied. This is a functional solution to adding a new language although really the code to add a new language should be centralised rather than the current distributed approach.

git-svn-id: https://svn.java.net/svn/javacc~svn/branches/version_61@484 68c0822f-77d1-6f10-b115-8a14d106ca4c