Added "LEGACY_EXCEPTION_HANDLING" option (defaults to false).
Generated (Java/GWT) code does now not throw Error (objects) by default when it encounters a runtime problem. Errors are only to be used when there is a failure that stops the JVM from continuing its operation. A failed parse is certainly an exception, but not an error (by the Java definition). As such, where Errors were previously thrown, JavaCC now throws RuntimeExceptions (also an unchecked exception).

If the user wishes to fall back to the previous behaviour then the used must provide the "LEGACY_EXCEPTION_HANDLING = true;" option in the options section at the top of the .jj file.

The CPP behaviour is unaffected by this flag.

git-svn-id: https://svn.java.net/svn/javacc~svn/branches/version_61@504 68c0822f-77d1-6f10-b115-8a14d106ca4c