Significant refactoring of Options settings and the heavy use of string literals throughout JavaCC. Now all javacc options (not jjtree yet) are registered with their type and their default value, and all options are now constance with user controlled option constants (significant reduction in use of string literals) having the USEROPTION__ prefix, and options set indirectly by javacc itself having the NONUSER_OPTION__ prefix.
The javacc command line help text now parses through the registry of options, and current set of default values to give the user accurate information that is guaranteed to be in-line with the code (previous approach was to manually log the options and defaults in static text within javacc which of course has become out of line with the source).

Replaced OUTPUT_LANGUAGE = "gwt" with OUTPUT_LANGUAGE = "java" (which is default anyway) and JAVA_TEMPLATE_TYPE = "modern";

This made a lot more sense as gwt is not a language, but rather a flavour of java compliance. The name of the option may change before relase in response to feedback, but the function (to alter the bahaviour of the code generation to decouple IO classes) will stay.

git-svn-id: https://svn.java.net/svn/javacc~svn/trunk@545 68c0822f-77d1-6f10-b115-8a14d106ca4c